requirements.txt
acres==0.5.0
adal==1.2.7
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
applicationinsights==0.11.10
appnope==0.1.4
argcomplete==3.5.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
azure-appconfiguration==1.7.1
azure-batch==15.0.0b2
azure-cli==2.74.0
azure-cli-core==2.74.0
azure-cli-telemetry==1.1.0
azure-common==1.1.28
azure-core==1.34.0
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==1.0.1
azure-identity==1.23.0
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.11.0b1
azure-keyvault-secrets==4.7.0
azure-keyvault-securitydomain==1.0.0b1
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-appconfiguration==3.1.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-batch==17.3.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==34.1.0
azure-mgmt-containerinstance==10.2.0b1
azure-mgmt-containerregistry==14.0.0
azure-mgmt-containerservice==36.0.0
azure-mgmt-core==1.5.0
azure-mgmt-cosmosdb==9.8.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datalake-store==1.1.0b1
azure-mgmt-datamigration==10.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0b3
azure-mgmt-imagebuilder==1.3.0
azure-mgmt-iotcentral==10.0.0b2
azure-mgmt-iothub==3.0.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==11.0.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==7.0.0b2
azure-mgmt-msi==7.0.0
azure-mgmt-mysqlflexibleservers==1.0.0b3
azure-mgmt-netapp==10.1.0
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-postgresqlflexibleservers==1.1.0b2
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b17
azure-mgmt-recoveryservices==3.0.0
azure-mgmt-recoveryservicesbackup==9.2.0
azure-mgmt-redhatopenshift==1.5.0
azure-mgmt-redis==14.5.0
azure-mgmt-resource==23.3.0
azure-mgmt-search==9.1.0
azure-mgmt-security==6.0.0
azure-mgmt-servicebus==8.2.1
azure-mgmt-servicefabric==2.1.0
azure-mgmt-servicefabricmanagedclusters==2.1.0b1
azure-mgmt-servicelinker==1.2.0b3
azure-mgmt-signalr==2.0.0b2
azure-mgmt-sql==4.0.0b21
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==23.0.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.3.1
azure-monitor-query==1.2.0
azure-multiapi-storage==1.4.0
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.20.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.7.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dotenv==0.9.9
et_xmlfile==2.0.0
etelemetry==0.3.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fabric==3.2.2
fastapi==0.115.14
fastembed==0.7.1
filelock==3.18.0
fitz==0.0.1.dev2
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
invoke==2.2.0
ipykernel==6.29.5
ipython==8.37.0
isodate==0.7.2
itsdangerous==2.2.0
javaproperties==0.5.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsondiff==2.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
knack==0.11.0
loguru==0.7.3
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
microsoft-security-utilities-secret-masker==1.0.0b4
mistralai==1.9.2
mmh3==5.1.0
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.6.3
narwhals==1.41.1
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
node==1.2.2
numpy==2.2.6
oauthlib==3.3.1
odict==1.9.0
onnxruntime==1.22.1
openai==1.84.0
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
paramiko==3.5.1
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pgvector==0.4.1
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.8
plumber==1.7
portalocker==2.10.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
prov==2.1.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
py-deviceid==0.1.1
py_rust_stemmers==0.1.5
pyarrow==20.0.0
PyAutoGUI==0.9.54
pycomposefile==0.0.32
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
PyGetWindow==0.0.9
PyGithub==1.59.1
Pygments==2.19.1
PyJWT==2.10.1
pymsalruntime==0.17.1
PyMsgBox==1.0.9
PyMuPDF==1.26.3
PyNaCl==1.5.0
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytweening==1.2.0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.0.0
qdrant-client==1.14.3
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.1
rubicon-objc==0.5.1
scikit-learn==1.7.1
scipy==1.15.3
scp==0.13.6
seaborn==0.13.2
semver==3.0.4
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sshtunnel==0.1.5
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.3.3
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
xlsxwriter==3.2.5
xmltodict==0.14.2
yarl==1.20.1
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.1
zope.event==5.1
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.proxy==6.1
